generators/requirements.txt
setuptools<68
pystac-client<0.6
pyyaml<7
redis<4
pygeofilter[backend-native]==0.2.0
click<9
requests<3
lxml<5
python-dateutil<3
swiftspec==0.0.2
mergedeep
structlog<22.0
pystac<1.8
vs-common @ git+https://gitlab.eox.at/vs/vs-common.git@0.5.0
harvester @ git+https://gitlab.eox.at/vs/vs.git@harvester-update-stacapi#egg=harvester&subdirectory=harvester